requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.14.0

# Audio Processing
sounddevice>=0.4.6
numpy>=1.24.0
scikit-learn>=1.3.0

# Text-to-Speech
pyttsx3>=2.90
piper-tts>=1.2.0
onnxruntime>=1.16.0

# Data Processing
pandas>=1.5.0
librosa>=0.10.0
soundfile>=0.12.0

# Utilities
accelerate>=0.20.0
protobuf>=3.20.0

# LLM Client
ollama

# ASR decoding improvements
pyctcdecode>=0.5.0
kenlm>=0.2.0
requests>=2.31.0

# Optional: For advanced TTS (if needed)
# gtts>=2.3.0
# pydub>=0.25.0
